[REG-1059] use new config methods for menuBot
